LayoutTests/platform/glib/imported/w3c/web-platform-tests/html/capability-delegation/delegate-fullscreen-request-popup-cross-origin.https.sub.tentative-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: Cannot request fullscreen without transient activation.
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: Cannot request fullscreen without transient activation.
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: Cannot request fullscreen without transient activation.
Verifies that element.requestFullscreen() calls from a cross-origin popup without user activation work if and only if the opener has user activation and it delegates the capability. https://wicg.github.io/capability-delegation/spec.html

PASS Fullscreen requests from a cross-origin popup fails without delegation from an opener with no user activation
PASS Fullscreen requests from a cross-origin popup fails without delegation from an opener with user activation
FAIL Fullscreen requests from a cross-origin popup succeeds with delegation from an opener with user activation assert_equals: expected "success" but got "failure"
